Fixes related to benchmarking refactoring (#131)
