+++ benchmarking/benchmark_loop/benchmark_main.py
@@ -4,11 +4,14 @@ import logging
 from argparse import ArgumentParser
 from tqdm import tqdm
 
-from blackbox_repository.tabulated_benchmark import BlackboxRepositoryBackend
-from benchmarks.benchmark_loop.baselines import methods
-from benchmarks.benchmark_loop.benchmark_definitions import benchmark_definitions
+from benchmarking.blackbox_repository.tabulated_benchmark import \
+    BlackboxRepositoryBackend
+from benchmarking.benchmark_loop.baselines import methods
+from benchmarking.benchmark_loop.benchmark_definitions import \
+    benchmark_definitions
 
-from syne_tune.backend.simulator_backend.simulator_callback import SimulatorCallback
+from syne_tune.backend.simulator_backend.simulator_callback import \
+    SimulatorCallback
 from syne_tune.stopping_criterion import StoppingCriterion
 from syne_tune.tuner import Tuner
 from coolname import generate_slug
