+++ docs/tutorials/basics/scripts/traincode_report_withcheckpointing.py
@@ -26,7 +26,7 @@ import argparse
 import logging
 
 from syne_tune.report import Reporter
-from benchmarks.checkpoint import resume_from_checkpointed_model, \
+from benchmarking.utils import resume_from_checkpointed_model, \
     checkpoint_model_at_rung_level, add_checkpointing_to_argparse, \
     pytorch_load_save_functions
