+++ examples/training_scripts/height_with_cost/train_height_with_cost.py
@@ -21,9 +21,8 @@ import math
 
 from syne_tune.report import Reporter
 from syne_tune.search_space import randint, add_to_argparse
-from benchmarking.utils.checkpoint import resume_from_checkpointed_model, \
-    checkpoint_model_at_rung_level, add_checkpointing_to_argparse
-from benchmarks.utils import parse_bool
+from benchmarking.utils import resume_from_checkpointed_model, \
+    checkpoint_model_at_rung_level, add_checkpointing_to_argparse, parse_bool
 
 
 _config_space = {
